# test: Optimizations