--- a/test/verify/check-shutdown-restart
+++ b/test/verify/check-shutdown-restart
@@ -25,20 +25,22 @@ import re
 import ast
 
 class TestShutdownRestart(MachineCase):
+    additional_machines = {
+        'machine2': { }
+    }
+
     def setUp(self):
         # we need a static ip for this test
         MachineCase.setUp(self, macaddr='52:54:00:9e:00:F3')
+
         self.machine.execute("hostnamectl set-hostname machine1")
-        self.machine2 = self.new_machine()
-        self.machine2.start()
-        self.machine2.wait_boot()
-        self.machine2.execute("hostnamectl set-hostname machine2")
+        self.machines['machine2'].execute("hostnamectl set-hostname machine2")
 
     def testBasic(self):
         m = self.machine
         b = self.browser
 
-        m2 = self.machine2
+        m2 = self.machines['machine2']
         b2 = self.new_browser(m2.address)
 
         # HACK -https://bugzilla.redhat.com/show_bug.cgi?id=1278287