--- a/test/verify/naughty-rhel-atomic/3455-tmp-read-only-4
+++ b/test/verify/naughty-rhel-atomic/3455-tmp-read-only-4
@@ -1,3 +1,3 @@
 Traceback (most recent call last):
-  File "check-multi-machine", line 314, in testFrameReload
+  File "check-multi-machine", line 331, in testFrameReload
     b.wait_text('#file-content', "1")