--- a/test/verify/naughty/4538-gss-continue-needed
+++ b/test/verify/naughty/4538-gss-continue-needed
@@ -1,3 +1,3 @@
 Traceback (most recent call last):
-  File "./verify/check-realms", line 232, in testNegotiate
+  File "./verify/check-realms", line 222, in testNegotiate
     self.assertIn("HTTP/1.1 200 OK", output)