blocked on #4740 

Boot images simultaneously, I still have a few occurrences to replace.
Results locally:
```
verify/check-multi-machine
# TESTS PASSED [470s on devmachine]
->
# TESTS PASSED [341s on devmachine]


verify/check-openshift
# TESTS PASSED [851s on devmachine]
->
# TESTS PASSED [646s on devmachine]
```