@dperpeet 2016-07-14
reference rhel-7 test without these changes: https://fedorapeople.org/groups/cockpit/logs/pull-4740-daf0eb4d-verify-rhel-7/log.html (on cockpit-7)