@dperpeet 2016-07-14
timing improvements locally:

verify/check-multi-machine
TESTS PASSED [470s on devmachine]
->
TESTS PASSED [341s on devmachine]


verify/check-openshift
TESTS PASSED [851s on devmachine]
->
TESTS PASSED [646s on devmachine]

./verify/check-realms
TESTS PASSED [285s on devmachine]
->
TESTS PASSED [200s on devmachine]


./verify/check-dashboard
TESTS PASSED [246s on devmachine]
->
TESTS PASSED [181s on devmachine]


./verify/check-multi-machine-key
TESTS PASSED [57s on devmachine]
->
TESTS PASSED [39s on devmachine]


./verify/check-shutdown-restart
TESTS PASSED [111s on devmachine]
->
TESTS PASSED [91s on devmachine]


./verify/check-subscriptions
TESTS PASSED [89s on devmachine]
->
TESTS PASSED [67s on devmachine]


./verify/check-multi-os
TESTS PASSED [57s on devmachine]
->
TESTS PASSED [35s on devmachine]