@dperpeet 2016-07-14
TESTS PASSED [1403s on cockpit-7] [logs](https://fedorapeople.org/groups/cockpit/logs/pull-4740-daf0eb4d-verify-rhel-7/log)
->
TESTS PASSED [1059s on cockpit-7] [logs](https://fedorapeople.org/groups/cockpit/logs/pull-4741-ecf1e10b-verify-rhel-7/log)