@dperpeet 2016-07-15
> I think the naughty file 3455 for rhel atomic needs to be updated.
> 
> I also saw a couple ip address not found failures that we had not seen in a long time, triggered those tests to see if it happens again.

Good catch. I clicked through all of them, I thought - apparently I missed that one. =)