@dperpeet 2016-07-18
addressed all above issues