@dperpeet 2016-07-18
> Seems good to me, only thing is a nit with the key names in check-dashboard. If you agree and I just fix when merging.

I can fix that, but apparently there are some testing failures. Will have a look.